format shard text
