+++ src/components/Filters/TransactionsFilters/ShardColumnFilters.tsx
@@ -4,8 +4,7 @@ import { useSelector } from 'react-redux';
 import { useSearchParams } from 'react-router-dom';
 
 import { SelectFilter } from 'components';
-import { getShardText } from 'helpers';
-import { useFetchShards, useIsSovereign } from 'hooks';
+import { useFetchShards, useIsSovereign, useGetShardText } from 'hooks';
 import { faFilter } from 'icons/regular';
 import { faFilter as faFilterSolid } from 'icons/solid';
 import { shardsSelector } from 'redux/selectors';
@@ -19,6 +18,7 @@ export const ShardColumnFilters = ({
   const stateShards = useSelector(shardsSelector);
   const [searchParams] = useSearchParams();
   const isSovereign = useIsSovereign();
+  const getShardText = useGetShardText();
   const { senderShard, receiverShard } = Object.fromEntries(searchParams);
 
   useFetchShards();
@@ -26,7 +26,7 @@ export const ShardColumnFilters = ({
   const selectShards = stateShards.map((shard) => {
     return {
       value: shard.shard.toString(),
-      label: getShardText(shard.shard.toString(), isSovereign)
+      label: getShardText(shard.shard.toString())
     };
   });
