+++ src/components/ShardSpan/ShardSpan.tsx
@@ -1,12 +1,11 @@
-import { getShardText } from 'helpers';
-import { useIsSovereign } from 'hooks';
+import { useGetShardText } from 'hooks';
 
 interface ShardSpanType {
   shard: number | string;
 }
 
 export const ShardSpan = ({ shard }: ShardSpanType) => {
-  const isSovereign = useIsSovereign();
+  const getShardText = useGetShardText();
 
-  return <span>{getShardText(shard, isSovereign)}</span>;
+  return <span>{getShardText(shard)}</span>;
 };
