+++ src/helpers/getValue/index.ts
@@ -7,7 +7,6 @@ export * from './getNodeStatusDisplay';
 export * from './getOperationDirection';
 export * from './getPercentage';
 export * from './getPrimaryColor';
-export * from './getShardText';
 export * from './getStringPlural';
 export * from './getSubdomainNetwork';
 export * from './getTotalTokenUsdValue';
